+++ src/IAMClient4D.Security.JWT.Verifiers.RSA.pas
@@ -93,13 +93,13 @@ implementation
 
 uses
   IAMClient4D.Exceptions,
-  IAMClient4D.Security.Crypto.LockBox3;
+  IAMClient4D.Security.Crypto.Factory;
 
 { TRSAJWTSignatureVerifier }
 
 constructor TRSAJWTSignatureVerifier.Create;
 begin
-  Create(TIAM4DLockBox3CryptoProvider.Create);
+  Create(TIAM4DCryptoProviderFactory.CreateProvider);
 end;
 
 constructor TRSAJWTSignatureVerifier.Create(const ACryptoProvider: IIAM4DCryptoProvider);